New corn test and a bunch of other adjustments
I don't know how my changes could equate to 2897 changes, but that is
what git hub is telling me.  I'll go ahead and commit this but please
tell me if the project opens for you and seems to work ok. Basically you
should see a bunch of tall corn stalks in the scene.